requirements.txt
botbuilder-core
botbuilder-schema
botbuilder-integration-aiohttp
python-dotenv
aiohttp